debian.mk: factor out and reuse code to create debian package
As a result now both the normal and the -nox packages have
the correct control files and pass lintian tests.

Note that we need a couple more commits to get a working
tree for creating debian packages.

While there also patch the Neubot makefile to avoid generating
the .pyc files, which slows down the process.