Repaired the conflict, inst, and reinst regression tests
Apparently those tests do work now, i.e., they do not "raise"
errors and they always show the notification window.

While there drop a sudo in front of each script in the Makefile,
to avoid creating the packages as root (lintian does not like that,
and I agree with it).

Also pass the current version number to the scripts, so we don't
need to hardcode the version in each script.

Also avoid generating the packages in each script, because the
makefile regress rule depends on the packages.

Also packages are now in the rootdir not inside dist.

While there NEXA -> Nexa in copyright headers.