regress/Debian/control/postinst: don't wait for notification window
Apparently on Ubuntu 12.04 there is no notification window.

I suppose this is because a neubot executable is not actually
installed by this regression test.

Anyway at the level of abstraction of this regression test
it suffices to check that postinst creates the file that
will trigger the notification window when Neubot is fully
installed (e.g., by other regression tests).

In fact there are other regression tests in which we check
that the notification window appears.

Therefore I lose nothing, by removing the code to wait for
the window from this regression test.

Probably I'm also making the whole regress suite a bit faster.